fix code

+++ src/Model/Ship.cs
@@ -1,9 +1,7 @@
-
 using Microsoft.VisualBasic;
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Data;
 using System.Diagnostics;
 /// <summary>
 /// A Ship has all the details about itself. For example the shipname,
@@ -80,7 +78,7 @@ public class Ship
         _tiles = new List<Tile>();
 
         //gets the ship size from the enumarator
-        _sizeOfShip = _shipName;
+        _sizeOfShip = (int)_shipName;
     }
 
     /// <summary>
@@ -133,10 +131,3 @@ public class Ship
         _direction = direction;
     }
 }
-
-//=======================================================
-//Service provided by Telerik (www.telerik.com)
-//Conversion powered by NRefactory.
-//Twitter: @telerik
-//Facebook: facebook.com/telerik
-//=======================================================
